# open_api/openapi.yaml
openapi: 3.0.3
info:
  title: FormulOne
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: testing server
...
paths:
  /api/auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: Login
      requestBody:
        description: Login
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Invalid login or password

        '500':
          description: Status internal server error
  /api/auth/register:
    post:
      tags:
        - auth
      summary: Register
      description: Register
      requestBody:
        description: Register
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '500':
          description: Status internal server error
...
  /api/drivers:
    get:
      tags:
        - drivers
      summary: Get all drivers
      description: Returns all drivers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Driver'
        '401':
          description: Unauthorized user
        '500':
          description: Status internal server error

    post:
      tags:
        - drivers
      summary: create driver
      description: ''
      requestBody:
        description: Create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Driver'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '500':
          description: Status internal server error
  /api/drivers/{id}:
    get:
      tags:
        - drivers
      summary: Get driver
      description: Return driver by id
      parameters:
        - name: id
          in: path
          description: ID of driver
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    put:
      tags:
        - drivers
      summary: Update driver
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of driver
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update driver
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Driver'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    delete:
      tags:
        - drivers
      summary: Delete driver
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of driver
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
...
  /api/grandprix:
    get:
      tags:
        - grand_prix
      summary: Get all grand prix
      description: Returns all grand prix
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrandPrixResponse'
        '500':
          description: Status internal server error

    post:
      tags:
        - grand_prix
      summary: Create grand prix
      description: ''
      requestBody:
        description: Create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrandPrix'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '500':
          description: Status internal server error
  /api/grandprix/{id}:
    get:
      tags:
        - grand_prix
      summary: Get grand prix
      description: Return grand prix by id
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrandPrixResponse'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    put:
      tags:
        - grand_prix
      summary: Update grand prix
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update grand prix
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrandPrix'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    delete:
      tags:
        - grand_prix
      summary: Delete grand prix
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of grand prix
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
...
  /api/grandprix/featured:
    get:
      tags:
        - featured
      summary: Get all current user's featured grand prix
      description: Returns all featured grand prix
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrandPrixResponse'
        '401':
          description: Unauthorized user
        '500':
          description: Status internal server error
  /api/grandprix/{id}/featured:
      post:
        tags:
          - featured
        summary: Add feature tag to grand prix
        description: ''
        parameters:
          - name: id
            in: path
            description: ID of grand prix
            required: true
            schema:
              type: integer
              format: int64
        responses:
          '200':
            description: Successful operation
          '401':
            description: Unauthorized user
          '404':
            description: Id doesnt exist
          '500':
            description: Status internal server error
      delete:
        tags:
          - featured
        summary: Delete feature tag on grand prix
        description: Delete feature tag
        parameters:
          - name: id
            in: path
            description: ID of grand prix
            required: true
            schema:
              type: integer
              format: int64
        responses:
          '200':
            description: Successful operation
          '401':
            description: Unauthorized user
          '404':
            description: Id doesnt exist
          '500':
            description: Status internal server error
components:
  schemas:
    Auth:
      type: object
      properties:
        login:
          type: string
          example: user
        password:
          type: string
          example: password
    Driver:
      type: object
      properties:
        driver_name:
          type: string
          example: Maaax
        driver_country:
          type: string
          example: Turkey
        driver_birth_date:
          type: integer
          example: 11.19.1991
...
    GrandPrix:
      type: object
      properties:
        gp_id:
          type: integer
          format: int64
          example: 121
        gp_season:
          type: integer
          format: int64
          example: 2020
        gp_name:
          type: string
          example: Monaco GP
        gp_date_num:
          type: integer
          format: int64
          example: 23
        gp_month:
          type: string
          example: may
        gp_place:
          type: string
          example: Hungary
        gp_track_id:
          type: integer
          format: int64
          example: 2
    GrandPrixResponse:
      type: object
      properties:
        gp_id:
          type: integer
          format: int64
          example: 121
        gp_season:
          type: integer
          format: int64
          example: 2020
        gp_name:
          type: string
          example: Monaco GP
        gp_date_num:
          type: integer
          format: int64
          example: 23
        gp_month:
          type: string
          example: may
        gp_place:
          type: string
          example: Hungary
        gp_track_id:
          type: integer
          format: int64
          example: 2
        gp_featured:
          type: boolean
          example: true
...